+++ app/src/main/java/com/example/swapit/data/datasource/RemoteUserDataSource.kt
@@ -1,6 +1,6 @@
 package com.example.swapit.data.datasource
 
-import com.example.swapit.data.datasource.remote.service.UserResponse
+import com.example.swapit.data.datasource.remote.dto.response.user.UserResponse
 import com.example.swapit.data.datasource.remote.service.UserService
 
 class RemoteUserDataSource(
