+++ app/src/main/java/com/example/swapit/domain/model/user/UserInfo.kt
@@ -1,0 +1,10 @@
+package com.example.swapit.domain.model.user
+
+data class UserInfo(
+    val id: Long,
+    val nickname: String,
+    val profileImageUrl: String,
+    val email: String,
+    val swapStats: UserSwapStats,
+    val reviews: List<Review>,
+)
