+++ app/src/main/java/com/example/swapit/domain/model/user/UserSwapStats.kt
@@ -1,0 +1,7 @@
+package com.example.swapit.domain.model.user
+
+data class UserSwapStats(
+    val totalGoodsCount: Long,
+    val completedSwapCount: Long,
+    val ratingAverage: Double,
+)
